+++ .gitignore
@@ -1,0 +1,3 @@
+/node-modules
+node_modules
+yarn.lock
